[Chore] #18 - 네트워크 폴더링 수정
