+++ 30th-SOPT-iOS-assignment/30th-SOPT-iOS-assignment/30th-SOPT-iOS-assignment/Data/Network/Essentials/NetworkResult.swift
